feat(mm-next/photography): add ArrowDown button click handler
